[lexer.c] : Add real numbers type and also lex them
change scan_number function, it will return double , and will set
token type, we have to check token type after call to use of it's value

+++ src/lexer/lexer.h
@@ -113,6 +113,6 @@ private void scan_ident();
 /*
  *	@brief : scan number from lexer , number can be decimal or octal or ....
  */
-private int scan_number(char);
+private double scan_number(char);
 
 #endif
